Create a results page with links to analysis files

create_links_page automates the curation of all the analyses for a project.

create_links_page(output = "results.Rmd", sort_method = "filename",
  path = ".", exclude_rmd = NULL, datatable = FALSE)

Arguments

output

The basename of the results file (default: "results.Rmd")

sort_method

The method for sorting the files. Options include "filename" (default), "title", "date", and "date reverse" (see Details for explanation).

path

By default the function assumes the current working directory is within the project. If this is not true, you'll need to provide the path to the project directory.

exclude_rmd

character vector of R Markdown files (basename only) to exclude from the results page

datatable

Display the list of results using DT::datatable (not yet implemented)

Value

Invisibly returns a data.frame containing information on the R Markdown files.

Details

By default, this creates the file "results.Rmd" in the analysis/ subdirectory. If you decide to change this name, be sure to update the navigation bar settings in _site.yml.

There are currently 4 methods available for sorting the files. "filename" (the default) sorts by filename, "title" sorts by title (if the R Markdown file does not contain a title in the YAML, the filename is used), "date" sorts in chronological order, and "date reverse" sorts in reverse chronological order.

Examples

## Not run: ------------------------------------
# create_links_page()
## ---------------------------------------------