post_knit function for workflowr

This is the post_knit function that wflow_html passes to the function output_format from the package rmarkdown. For advanced usage only.

wflow_post_knit(metadata, input_file, runtime, encoding, ...)

Arguments

metadata

The metadata specified in the YAML header of the R Markdown file

input_file

Name of R Markdown file

runtime

The runtime target for rendering. The static option produces output intended for static files; shiny produces output suitable for use in a Shiny document (see run). The default, auto, allows the runtime target specified in the YAML metadata to take precedence, and renders for a static runtime target otherwise.

encoding

Ignored. The encoding is always assumed to be UTF-8.

...

arguments passed to the post_knit function of rmarkdown::html_document

Details

If you'd like to combine workflowr with another R Markdown output format, you may need to use wflow_post_knit. This function fixes the path to the R Markdown file (which is manipulated by wflow_pre_knit).

See also