pre_processor function for workflowr

This is the pre_processor function that wflow_html passes to the function output_format from the package rmarkdown. For advanced usage only.

wflow_pre_processor(metadata, input_file, runtime, knit_meta, files_dir,
  output_dir)

Arguments

metadata
The metadata specified in the YAML header of the R Markdown file

input_file
Name of Markdown file created by knitr::knit to be passed to pandoc

runtime
The runtime target for rendering. The static option produces output intended for static files; shiny produces output suitable for use in a Shiny document (see run). The default, auto, allows the runtime target specified in the YAML metadata to take precedence, and renders for a static runtime target otherwise.

knit_meta
(This option is reserved for expert use.) Metadata generated by knitr.

files_dir
Directory for saving intermediate files

output_dir
The output directory for the rendered output_file. This allows for a choice of an alternate directory to which the output file should be written (the default output directory of that of the input file). If a path is provided with a filename in output_file the directory specified here will take precedence. Please note that any directory path provided will create any necessary directories if they do not exist.

Details

If you'd like to combine workflowr with another R Markdown output format, you may need to use wflow_pre_processor. This function only has minor effects on the style of the resulting HTML page, and is not essential for using workflowr.

See also