Rename a workflowr project

If you want to rename an existing workflowr project, use wflow_rename_proj to update the name throughout all the project files.

wflow_rename_proj(
  name,
  rproj = TRUE,
  remote = TRUE,
  navbar = TRUE,
  readme = TRUE,
  commit = TRUE,
  directory = TRUE,
  project = "."
)

Arguments

name
character. The new name for the workflowr project.

rproj
logical (default: TRUE). Rename the RStudio Project file.

remote
logical (default: TRUE). Rename the remote URL.

navbar
logical (default: TRUE). Rename the navbar title.

readme
logical (default: TRUE). Rename the README title.

commit
logical (default: TRUE). Commit the changes to Git.

directory
logical (default: TRUE). Rename the project directory.

project
character (default: ".") By default the function assumes the current working directory is within the project. If this is not true, you'll need to provide the path to the project directory.

Value

Invisibly returns the path to the project directory

Details

wflow_rename_proj performs the following steps and then commits the changes:

Rename RStudio Project file (.Rproj)

Update URL of remote repository (see wflow_git_remote)

Update project name in the navigation bar (defined in _site.yml)

Update title of README file

Rename the project directory itself

After renaming the project with wflow_rename_proj, you should republish the R Markdown files with wflow_publish(republish = TRUE). Also, you should go to the settings of your Git repository on the online Git hosting platform to change its name.

See also

Examples

if (FALSE) {

wflow_rename_proj("new-project-name")
}