Start a new workflowr project.

wflow_start adds the necesary files for a workflowr project. The default is to create a new directory, but it is also possible to add the files to an already existing project.

wflow_start(name, directory, git = TRUE, existing = FALSE,
  overwrite = FALSE)

Arguments

name

character. Project name, e.g. "My Project"

directory

character. The directory for the project, e.g. "~/new-project". By default the directory will be created (existing = FALSE.

git

logical (default: TRUE). Should Git be used for version control? If directory is a new Git repository, wflow_start will initialize the repository and make an initial commit. If directory is already a Git repository, wflow_start will make an additional commit. In both cases, only files needed for the workflowr project will be included in the commit.

existing

logical (default: FALSE). Indicate if the specified directory already exists. The default prevents injecting the workflowr files into an unwanted location. Only set to TRUE if you wish to add the workflowr files to an existing project.

overwrite

logical (default: FALSE). Control whether to overwrite existing files. Only relevant if existing = TRUE. Passed to file.copy.

Details

This is the initial function that organizes the infrastructure to create a research website for your project. Note that while you do not need to use RStudio with workflowr, do not delete the Rproj file because it is required by other functions.

Examples

## Not run: ------------------------------------
# wflow_start("My Project", "path/to/new-project")
## ---------------------------------------------