Custom site generator for workflowr websites

wflow_site is a custom site generator to be used in combination with the R Markdown output format wflow_html.

wflow_site(input, encoding = getOption("encoding"), ...)

Arguments

input
character. The name of the website directory or a specific R Markdown file in the website directory.

encoding
character. The character encoding to use to read the file.

...
Placeholder for potential future use.

Details

Do not call the function wflow_site directly. Instead insert the line below directly into the YAML header of the file index.Rmd:

---
title: "Home"
site: workflowr::wflow_site
output:
  workflowr::wflow_html:
    toc: false
---

Then you can build the website by running render_site in the R console or clicking the Knit button in RStudio.

If you receive an error when using the RStudio Knit button (the error is about an unused argument), make sure the Knit Directory is set to Document Directory (you can set this with the dropdown menu next to the Knit button).

See also