workflowr: A workflow template for creating a research website

The workflowr package helps you create a research website using R Markdown and Git.

Usage,NULL

Vignettes

Run browseVignettes("workflowr") to read the package vignettes locally. Alternatively you can read the documentation online at https://jdblischak.github.io/workflowr.

Main workflowr functions

wflow_start

Start a workflowr project.

wflow_build

Build the site to view locally.

wflow_publish

Publish analyses to include in the website.

wflow_status

Report status of analysis files.

Supporting workflowr functions

For further information on workflowr, see the help pages for these functions:

wflow_html

More technical details about how individual R Markdown files are converted to webpages, and how the rendering settings can be customized.

wflow_site

This help page explains how project-wide rendering settings can be customized in the _site.yml file.

Package options

The following package options affect the default behavior of the workflowr functions. To permanently set any of these options, add a call to the function options in the file .Rprofile at the root of your workflowr project. For example:

# Do not use Git executable
options(workflowr.sysgit = "")

workflowr.autosave

A logical indicating whether workflowr functions should automatically save files open in the RStudio editor before running. The default is TRUE. This requires RStudio 1.1.287 or later. Only files that have been previously saved are affected. In other words, unnamed files will be ignored.

workflowr.sysgit

The path to the system Git executable. This is occasionally used to increase the speed of Git operations performed by workflowr. By default it is set to the first Git executable on the search path. You can specify a path to a different Git executable. Alternatively you can disable this behavior entirely by setting it to the empty string "".

workflowr.view

A logical indicating whether workflowr functions should open webpages for viewing in the browser. The default is set to interactive (i.e. it is TRUE only if it is an interactive R session). This option is currently used by wflow_build, wflow_git_push, and wflow_publish.