workflowr: A workflow template for creating a research website

The workflowr package helps you create a research website using R Markdown and Git.

Vignettes

Run browseVignettes("workflowr") to read the package vignettes locally. Alternatively you can read the documentation online at https://jdblischak.github.io/workflowr.

Main workflowr functions

wflow_start - Start workflowr project

wflow_build - Build the site to view locally

wflow_publish - Publish analyses to include in the website

wflow_status - Report status of analysis files